Pull out room modules into their own files
