+++ lib/kalevala/world/room.ex
@@ -1,277 +1,3 @@
-defmodule Kalevala.World.Room.Context do
-  @moduledoc """
-  Context for performing work for an event in a room
-  """
-
-  @type t() :: %__MODULE__{}
-
-  defstruct [:data, assigns: %{}, characters: [], events: [], item_instances: [], lines: []]
-
-  defp push(context, to_pid, event = %Kalevala.Character.Conn.Event{}, _newline) do
-    Map.put(context, :lines, context.lines ++ [{to_pid, event}])
-  end
-
-  defp push(context, to_pid, data, newline) do
-    lines = %Kalevala.Character.Conn.Lines{
-      data: data,
-      newline: newline
-    }
-
-    Map.put(context, :lines, context.lines ++ [{to_pid, lines}])
-  end
-
-  @doc """
-  Render text back to a pid
-  """
-  def render(context, to_pid, view, template, assigns) do
-    assigns = Map.merge(context.assigns, assigns)
-    data = view.render(template, assigns)
-    push(context, to_pid, data, false)
-  end
-
-  @doc """
-  Render a prompt back to a pid
-  """
-  def prompt(context, to_pid, view, template, assigns) do
-    assigns = Map.merge(context.assigns, assigns)
-    data = view.render(template, assigns)
-    push(context, to_pid, data, true)
-  end
-
-  @doc """
-  Add to the assignment map on the context
-  """
-  def assign(context, key, value) do
-    assigns = Map.put(context.assigns, key, value)
-    Map.put(context, :assigns, assigns)
-  end
-
-  @doc """
-  Send an event back to a pid
-  """
-  def event(context, to_pid, from_pid, topic, data) do
-    event = %Kalevala.Event{
-      from_pid: from_pid,
-      topic: topic,
-      data: data
-    }
-
-    Map.put(context, :events, context.events ++ [{to_pid, event}])
-  end
-end
-
-defmodule Kalevala.World.Room.Item do
-  @moduledoc """
-  Handle item pickup
-  """
-
-  alias Kalevala.Event
-  alias Kalevala.Event.ItemDrop
-  alias Kalevala.Event.ItemPickUp
-
-  @doc """
-  Handle a request for dropping an item
-  """
-  def handle_drop_request({:abort, event, reason}, state, metadata) do
-    character = event.acting_character
-
-    event = %Event{
-      from_pid: self(),
-      topic: ItemDrop.Abort,
-      metadata: metadata,
-      data: %ItemDrop.Abort{
-        from: state.data.id,
-        item_instance: event.data.item_instance,
-        reason: reason
-      }
-    }
-
-    send(character.pid, event)
-
-    state
-  end
-
-  def handle_drop_request({:proceed, event, item_instance}, state, metadata) do
-    character = event.acting_character
-
-    event = %Event{
-      from_pid: self(),
-      topic: ItemDrop.Commit,
-      metadata: metadata,
-      data: %ItemDrop.Commit{
-        from: state.data.id,
-        item_instance: item_instance
-      }
-    }
-
-    send(character.pid, event)
-
-    add_item_instance(state, item_instance)
-  end
-
-  @doc """
-  Handle a request for picking up an item
-  """
-  def handle_pickup_request({:abort, event, reason}, state, metadata) do
-    character = event.acting_character
-
-    event = %Event{
-      from_pid: self(),
-      topic: ItemPickUp.Abort,
-      metadata: metadata,
-      data: %ItemPickUp.Abort{
-        item_name: event.data.item_name,
-        from: state.data.id,
-        reason: reason
-      }
-    }
-
-    send(character.pid, event)
-
-    state
-  end
-
-  def handle_pickup_request({:proceed, event, item_instance}, state, metadata) do
-    character = event.acting_character
-
-    event = %Event{
-      from_pid: self(),
-      topic: ItemPickUp.Commit,
-      metadata: metadata,
-      data: %ItemPickUp.Commit{
-        from: state.data.id,
-        item_name: event.data.item_name,
-        item_instance: item_instance
-      }
-    }
-
-    send(character.pid, event)
-
-    remove_item_instance(state, item_instance)
-  end
-
-  defp add_item_instance(state, item_instance) do
-    item_instances = [item_instance | state.private.item_instances]
-    private = Map.put(state.private, :item_instances, item_instances)
-    Map.put(state, :private, private)
-  end
-
-  defp remove_item_instance(state, item_instance) do
-    item_instances =
-      Enum.reject(state.private.item_instances, fn room_item_instance ->
-        room_item_instance.id == item_instance.id
-      end)
-
-    private = Map.put(state.private, :item_instances, item_instances)
-    Map.put(state, :private, private)
-  end
-end
-
-defmodule Kalevala.World.Room.Movement do
-  @moduledoc """
-  Handle room movement
-  """
-
-  alias Kalevala.Event
-  alias Kalevala.Event.Display
-  alias Kalevala.Event.Movement
-  alias Kalevala.Event.Movement.Voting
-  alias Kalevala.World.Zone
-
-  @doc """
-  Handle the movement request
-
-  Called after `Kalevala.World.Room.movement_request/2`.
-
-  - If an abort, forward to the character
-  - Otherwise, Forward to the zone
-  """
-  def handle_request({:abort, event, reason}, state, metadata) do
-    %{character: character} = event.data
-
-    event = %Event{
-      from_pid: self(),
-      topic: Voting,
-      metadata: metadata,
-      data: %Voting{
-        aborted: true,
-        character: character,
-        from: state.data.id,
-        exit_name: event.data.exit_name,
-        reason: reason
-      }
-    }
-
-    send(character.pid, Voting.abort(event))
-  end
-
-  def handle_request({:proceed, event, room_exit}, state, metadata) do
-    %{character: character} = event.data
-
-    event = %Event{
-      from_pid: self(),
-      topic: Voting,
-      metadata: metadata,
-      data: %Voting{
-        character: character,
-        from: state.data.id,
-        to: room_exit.end_room_id,
-        exit_name: room_exit.exit_name
-      }
-    }
-
-    Zone.global_name(state.data.zone_id)
-    |> GenServer.whereis()
-    |> send(event)
-  end
-
-  @doc """
-  Handle the movement event
-  """
-  def handle_event(state, event = %Event{topic: Movement, data: %{direction: :to}}) do
-    state
-    |> broadcast(event)
-    |> append_character(event)
-  end
-
-  def handle_event(state, event = %Event{topic: Movement, data: %{direction: :from}}) do
-    state
-    |> reject_character(event)
-    |> broadcast(event)
-  end
-
-  @doc """
-  Broadcast the event to characters in the room
-  """
-  def broadcast(state, event) do
-    lines = %Kalevala.Character.Conn.Lines{data: event.data.reason, newline: true}
-    display_event = %Display{lines: [lines]}
-
-    Enum.each(state.private.characters, fn character ->
-      send(character.pid, display_event)
-    end)
-
-    state
-  end
-
-  defp append_character(state, event) do
-    characters = [event.data.character | state.private.characters]
-    private = Map.put(state.private, :characters, characters)
-
-    Map.put(state, :private, private)
-  end
-
-  defp reject_character(state, event) do
-    characters =
-      Enum.reject(state.private.characters, fn character ->
-        character.id == event.data.character.id
-      end)
-
-    private = Map.put(state.private, :characters, characters)
-    Map.put(state, :private, private)
-  end
-end
-
 defmodule Kalevala.World.Room.Private do
   @moduledoc """
   Store private information for a room, e.g. characters in the room
@@ -301,9 +27,8 @@ defmodule Kalevala.World.Room do
   alias Kalevala.Event.Message
   alias Kalevala.World
   alias Kalevala.World.Room.Context
+  alias Kalevala.World.Room.Events
   alias Kalevala.World.Room.Exit
-  alias Kalevala.World.Room.Item
-  alias Kalevala.World.Room.Movement
   alias Kalevala.World.Room.Private
 
   defstruct [
@@ -470,12 +195,10 @@ defmodule Kalevala.World.Room do
   def handle_call(event = %Event{topic: Event.Movement.Voting}, _from, state) do
     {context, event} =
       state
-      |> new_context()
+      |> Context.new()
       |> state.callback_module.confirm_movement(event)
 
-    context
-    |> send_lines()
-    |> send_events()
+    Context.handle_context(context)
 
     state = Map.put(state, :data, context.data)
 
@@ -483,132 +206,18 @@ defmodule Kalevala.World.Room do
   end
 
   @impl true
-  # Forward movement requests to the zone to handle
-  def handle_info(event = %Event{topic: Event.Movement.Request}, state) do
-    room_exit =
-      Enum.find(state.data.exits, fn exit ->
-        exit.exit_name == event.data.exit_name
-      end)
-
-    new_context(state)
-    |> state.callback_module.movement_request(event, room_exit)
-    |> Movement.handle_request(state, event.metadata)
-
-    {:noreply, state}
-  end
-
-  def handle_info(event = %Event{topic: Event.Movement}, state) do
-    case event.data.room_id == state.data.id do
-      true ->
-        state = Movement.handle_event(state, event)
-        {:noreply, state}
-
-      false ->
-        global_name(event.data.room_id)
-        |> GenServer.whereis()
-        |> send(event)
-
-        {:noreply, state}
-    end
-  end
-
-  def handle_info(event = %Event{topic: Event.ItemDrop.Request}, state) do
-    %{item_instance: item_instance} = event.data
-
-    state =
-      new_context(state)
-      |> state.callback_module.item_request_drop(event, item_instance)
-      |> Item.handle_drop_request(state, event.metadata)
-
-    {:noreply, state}
-  end
-
-  def handle_info(event = %Event{topic: Event.ItemPickUp.Request}, state) do
-    %{item_name: item_name} = event.data
-
-    item_instance =
-      Enum.find(state.private.item_instances, fn item_instance ->
-        item = state.callback_module.load_item(item_instance)
-        item.callback_module.matches?(item, item_name)
-      end)
-
-    state =
-      new_context(state)
-      |> state.callback_module.item_request_pickup(event, item_instance)
-      |> Item.handle_pickup_request(state, event.metadata)
-
-    {:noreply, state}
-  end
-
   def handle_info(event = %Event{}, state) do
-    context =
-      new_context(state)
-      |> state.callback_module.event(event)
-      |> handle_context()
-
-    state = Map.put(state, :data, context.data)
-
-    {:noreply, state}
+    Events.handle_event(event, state)
   end
 
   def handle_info(message = %Message{}, state) do
     context =
-      new_context(state)
+      Context.new(state)
       |> state.callback_module.event(message)
-      |> handle_context()
+      |> Context.handle_context()
 
     state = Map.put(state, :data, context.data)
 
     {:noreply, state}
-  end
-
-  defp new_context(state) do
-    item_instances =
-      Enum.map(state.private.item_instances, fn item_instance ->
-        meta = item_instance.callback_module.trim_meta(item_instance)
-        %{item_instance | meta: meta}
-      end)
-
-    characters =
-      Enum.map(state.private.characters, fn character ->
-        %{character | inventory: []}
-      end)
-
-    %Context{
-      data: state.data,
-      characters: characters,
-      item_instances: item_instances
-    }
-  end
-
-  defp handle_context(context) do
-    context
-    |> send_lines()
-    |> send_events()
-  end
-
-  defp send_lines(context) do
-    context.lines
-    |> Enum.group_by(
-      fn {to_pid, _line} ->
-        to_pid
-      end,
-      fn {_to_pid, line} ->
-        line
-      end
-    )
-    |> Enum.each(fn {to_pid, lines} ->
-      send(to_pid, %Event.Display{lines: lines})
-    end)
-
-    context
-  end
-
-  defp send_events(context) do
-    Enum.each(context.events, fn {to_pid, event} ->
-      send(to_pid, event)
-    end)
-
-    context
   end
 end
